mysite 완성

+++ board/views.py
@@ -9,6 +9,12 @@ def index(request):
     context = {'board_list': board_list}
     return render(request, 'board/list.html', context)
 
+#게시판 찾기
+def search(request):
+    board_seach = Board.objects.filter(title__contains=request.POST['kwd'])
+    context = {'board_list': board_seach}
+    return render(request, 'board/list.html', context)
+
 #게시판 등록 화면
 def writeform(request):
     return render(request, 'board/write.html')
@@ -26,9 +32,18 @@ def write(request):
 
 #게시물 detail
 def view(request):
+    #조회
     board_view = Board.objects.filter(id=request.GET['boardid'])
     board_view = model_to_dict(board_view[0])
+
+    #조회수+1
+    board_hit = Board.objects.get(id=request.GET['boardid'])
+    board_hit.hit = int(board_hit.hit)+1
+    board_hit.save()
+
+
     context = {'board_view':board_view}
+
     return render(request, 'board/view.html', context)
 
 #게시물 수정
@@ -39,6 +54,9 @@ def modifyform(request):
     return render(request, 'board/modify.html',context)
 
 def modify(request):
+    # if request.session['authuser'] is None :
+    #     return HttpResponseRedirect('/error')
+    # else:
     modify_data = Board.objects.get(id=request.POST['boardid'])
     modify_data.title=request.POST['title']
     modify_data.content=request.POST['content']
@@ -48,5 +66,9 @@ def modify(request):
 
 #게시물 delete TX
 def delete(request):
-    Board.objects.filter(id=request.GET['boardid']).delete()
-    return HttpResponseRedirect('/board')
+    try:
+        if request.session['authuser']['id'] is not None :
+            Board.objects.filter(id=request.GET['boardid']).delete()
+            return HttpResponseRedirect('/board')
+    except:
+        return HttpResponseRedirect('/error')
